[REFACTOR]: File name changed based on code review
